Document traits changed to AbstractDocument's
