+++ composer.json
@@ -13,7 +13,7 @@
     "require": {
         "php": ">=5.4",
         "symfony/symfony": "~2.6",
-        "ongr/elasticsearch-bundle": "~0.1",
+        "ongr/elasticsearch-bundle": "~0.6.0",
         "ongr/router-bundle": "~0.1"
     },
     "require-dev": {
